// js/packages/phoenix-client/src/__generated__/api/v1.d.ts
export interface paths {
  "/v1/datasets": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** List datasets */
    get: operations["listDatasets"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/v1/datasets/{id}": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** Get dataset by ID */
    get: operations["getDataset"];
    put?: never;
    post?: never;
    /** Delete dataset by ID */
    delete: operations["deleteDatasetById"];
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/v1/datasets/{id}/versions": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** List dataset versions */
    get: operations["listDatasetVersionsByDatasetId"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/v1/datasets/upload": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get?: never;
    put?: never;
    /** Upload dataset from JSON, CSV, or PyArrow */
    post: operations["uploadDataset"];
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/v1/datasets/{id}/examples": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** Get examples from a dataset */
    get: operations["getDatasetExamples"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/v1/datasets/{id}/csv": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** Download dataset examples as CSV file */
    get: operations["getDatasetCsv"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/v1/datasets/{id}/jsonl/openai_ft": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** Download dataset examples as OpenAI fine-tuning JSONL file */
    get: operations["getDatasetJSONLOpenAIFineTuning"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/v1/datasets/{id}/jsonl/openai_evals": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** Download dataset examples as OpenAI evals JSONL file */
    get: operations["getDatasetJSONLOpenAIEvals"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/v1/datasets/{dataset_id}/experiments": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** List experiments by dataset */
    get: operations["listExperiments"];
    put?: never;
    /** Create experiment on a dataset */
    post: operations["createExperiment"];
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/v1/experiments/{experiment_id}": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** Get experiment by ID */
    get: operations["getExperiment"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/v1/span_annotations": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get?: never;
    put?: never;
    /** Create or update span annotations */
    post: operations["annotateSpans"];
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/v1/evaluations": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** Get span, trace, or document evaluations from a project */
    get: operations["getEvaluations"];
    put?: never;
    /** Add span, trace, or document evaluations */
    post: operations["addEvaluations"];
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
}
export type webhooks = Record<string, never>;
export interface components {
  schemas: {
    /** AnnotateSpansRequestBody */
    AnnotateSpansRequestBody: {
      /** Data */
      data: components["schemas"]["SpanAnnotation"][];
    };
// ...
    SpanAnnotation: {
      /**
       * Span Id
       * @description OpenTelemetry Span ID (hex format w/o 0x prefix)
       */
      span_id: string;
      /**
       * Name
       * @description The name of the annotation
       */
      name: string;
      /**
       * Annotator Kind
       * @description The kind of annotator used for the annotation
       * @enum {string}
       */
      annotator_kind: "LLM" | "HUMAN";
      /** @description The result of the annotation */
      result?: components["schemas"]["SpanAnnotationResult"] | null;
      /**
       * Metadata
       * @description Metadata for the annotation
       */
      metadata?: Record<string, never> | null;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
  listDatasets: {
    parameters: {
      query?: {
        /** @description Cursor for pagination */
        cursor?: string | null;
        /** @description An optional dataset name to filter by */
        name?: string | null;
        /** @description The max number of datasets to return at a time. */
        limit?: number;
      };
      header?: never;
      path?: never;
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description Successful Response */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          "application/json": components["schemas"]["ListDatasetsResponseBody"];
        };
      };
      /** @description Forbidden */
      403: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          "text/plain": string;
        };
      };
      /** @description Unprocessable Entity */
      422: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          "text/plain": string;
        };
      };
    };
  };
// ...
}